Workaround for cra compilerOptions.paths
CRA does not allow us to set paths in compilerOptions, which is required
for our project in order to be able to import the abi file for our smart
contract. Therefore we create a new json file and save the paths
directive there, and then tell tsconfig to extend itself by referencing
the json file.